Add ability to download X100F configuration backup.
Fuji have their "Aquire" software which allows you to backup and restore
configuration data. However it only works on Windows and OS X :(

After some sniffing of the USB traffic it seems, at least for the backup
part, to be using standard(ish) ObjectGet request but with a handle of
`0` (which is undefined according to the spec)

The catch is that when performing the GetObjectInfo request the ptpy
fails to parse the response as it contains a null date.

Lets create an extension to house as much of our custom/hacky code as
possible :)